requirements.txt
fairseq
transformers
datasets
gradio
httpx==0.23.0
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
pyyaml
# *** others ***
timm
ftfy==6.0.3
tensorboardX==2.4.1
pycocotools==2.0.4
pytorch_lightning
einops
datasets
rouge_score
soundfile
editdistance
librosa
python-Levenshtein
zhconv
pypinyin==0.47.1
g2p_en